+++ tests/functional/test_functional.py
@@ -22,9 +22,14 @@ def test_expected_outcome(feature_file: Path, expected_output: Path, capsys) -> 
     linter = GherkinLinter(feature_file)
     linter.run()
     captured = capsys.readouterr()
+    output = _patch_filename(feature_file, captured.out)
+    assert output.strip() == expected_output.read_text("utf8").strip()
+
+
+def _patch_filename(feature_file: Path, capture: str) -> str:
     # replace full file paths with only the filename to be independent from the
     # runtime environment
     full_file_path = str(feature_file.absolute())
     filename_only = feature_file.name
-    output = captured.out.replace(full_file_path, filename_only)
-    assert output.strip() == expected_output.read_text("utf8").strip()
+    output = capture.replace(full_file_path, filename_only)
+    return output